Water & Light & Texturfix ... need to add depth to water

+++ Core/Graphics/BlockContext.hpp
@@ -8,15 +8,15 @@ namespace Core {
 	namespace Graphics {
 		class BlockContext : public PrimitiveContext {
 		private:
-			ivec3 *mesh, *normals;
+			vec3 *mesh, *normals;
 			map<TextureID, vec2*> uvs;
 			vec4 center;
 			Texture texture;
 		public:
 			BlockContext();
 			~BlockContext();
-			virtual ivec3* getMesh() const { return mesh; }
-			virtual ivec3* getNormals() const { return normals; }
+			virtual vec3* getMesh() const { return mesh; }
+			virtual vec3* getNormals() const { return normals; }
 			virtual vec2* getUVs(TextureID tid) { return uvs[tid]; }
 			virtual Texture* getTexture() { return &texture; }
 			virtual vec4* getCenter() { return &center; }
