+++ Core/Graphics/SkyBlockContext.hpp
@@ -6,14 +6,14 @@ namespace Core {
 	namespace Graphics {
 		class SkyBlockContext : public PrimitiveContext {
 		private:
-			ivec3 *mesh, *normals;
+			vec3 *mesh, *normals;
 			vec2 *uvs;
 			Texture texture;
 		public:
 			SkyBlockContext();
 			~SkyBlockContext();
-			virtual ivec3* getMesh() const { return mesh; }
-			virtual ivec3* getNormals() const { return normals; }
+			virtual vec3* getMesh() const { return mesh; }
+			virtual vec3* getNormals() const { return normals; }
 			virtual vec2* getUVs(TextureID tid) { return uvs; }
 			virtual Texture* getTexture() { return &texture; }
 			virtual vec4* getCenter() { return nullptr; }
